Add provider details page
I had completely forgotten to add the provider details page. The page is used on the queue page to help determine which provider to use.

The provider.html is a pure copy-paste from the previous project, and should probably be refactored at some point.
Besides adding the page the necessary rest functions has been added as well.