Limit select from jobdiag to the previous 30 days
It turns out that some of the older production databases contains A LOT of rows in jobdiag, so many in fact that takes more than a minute to load. Plus some of the messages are more than two years old and therefor not really relevant anymore.
By limiting the select to only get messages less than 30 days old it improves performance plus it increases the usefulness of the information shown.