Make record summary page a bit more user friendly
It has always been a requirement that it should be possible to sort the
columns as a user. Now that has finally been implemented.
In addition the reload button has been moved to above the table, so you
no longer have to scroll to the bottom of the page every time.