IMS queue type
IMS queue type is needed for putting records on queue which doesn't have
a real record. The existing queue types are meant for corepo which should
only have entries for real records (local or enrichment) and not entries
where there is holdings but no record.
For IMS it is holding that is interesting and not the records.